[DateLoader] moving default_collate to be publicly accessible
[ghstack-poisoned]
